feat: when user deletes a file from their web app, this is also deleted from cloudinary (#57)
* feat: apply prisma migrate to modified schema for file

display name is unique (for a given user, in a given folder) but there is a 'name' that is in the background. this name will be used to call cloudinary destroy

* fix(filesController): import statements

* feat(filesController): when uploading a file, the unique 'name' field (unique across the universe of files) is what is generated with helper. this is used for the display name only if there is another file of the same name, in the same folder of the same user

* refactor(views): file's 'name's are actually their display names

* feat: when user deletes a file from their web app, it is also deleted in cloudinary

also do try/catch to catch any errors and be informative. also render the error page if the file cannot be found, or if it is already been deleted

+++ src/controllers/filesController.ts
@@ -2,8 +2,11 @@ import type { Request, Response } from "express";
 import { cloudinary } from "@/config/cloudinary.js";
 import { upload } from "@/lib/multer.js";
 import { prisma } from "@/lib/prisma.js";
-import { generateUniqueFilename, getFileExtension } from "@/utils/helpers.js";
-import { getCloudinaryDownloadLink } from "../utils/helpers";
+import {
+	generateUniqueFilename,
+	getCloudinaryDownloadLink,
+	getFileExtension,
+} from "@/utils/helpers.js";
 
 const getAllowedFileTypesForUpload = (): string => {
 	const MS_WORD_FILE_TYPES = [
@@ -49,7 +52,7 @@ export const fileDetailsGet = async (req: Request, res: Response) => {
 				"Cannot find file. Please return to the previous page to see all files.",
 		});
 
-	res.render("pages/fileDetails", { title: file.name, file });
+	res.render("pages/fileDetails", { title: file.displayName, file });
 };
 
 export const uploadFileGet = async (req: Request, res: Response) => {
@@ -114,7 +117,7 @@ export const uploadFilePost = [
 
 		const filesWithSameOriginalNameInFolder = await prisma.file.findMany({
 			where: {
-				name: fileForUpload.originalname,
+				displayName: fileForUpload.originalname,
 				userId: user.id,
 				folderId: Number(folderIdToAddFile),
 			},
@@ -124,7 +127,8 @@ export const uploadFilePost = [
 
 		const newFile = await prisma.file.create({
 			data: {
-				name: hasDuplicateOriginalFilenameInFolder
+				name: uniqueDisplayName,
+				displayName: hasDuplicateOriginalFilenameInFolder
 					? uniqueDisplayName
 					: fileForUpload.originalname,
 				sizeInKb: fileForUpload.size,
@@ -149,12 +153,37 @@ export const fileDelete = async (req: Request, res: Response) => {
 
 	const { id: fileId } = req.params;
 
-	const deleteFile = await prisma.file.delete({
+	const file = await prisma.file.findUnique({
+		where: { id: Number(fileId) },
+	});
+	if (!file)
+		return res.status(404).render("pages/error", {
+			statusCode: 404,
+			errorMessage: "File has already been deleted.",
+		});
+
+	const _deleteFile = await prisma.file.delete({
 		where: {
-			id: Number(fileId),
+			id: file.id,
 			userId: user.id,
 		},
 	});
 
-	res.redirect(`/folders/${deleteFile.folderId}`);
+	let cloudinaryDeleteResult = null;
+	try {
+		cloudinaryDeleteResult = await cloudinary.uploader.destroy(file.name);
+	} catch (err) {
+		console.error(err);
+		if (err instanceof Error) throw new Error(err.message);
+		throw new Error(`Error when deleting file from cloud: ${err}`);
+	}
+
+	if (cloudinaryDeleteResult.result === "ok")
+		res.redirect(`/folders/${file.folderId}`);
+	else
+		res.status(404).render("pages/error", {
+			statusCode: 404,
+			errorMessage:
+				"An error occurred when trying to delete that file from the cloud. It's possible that the file has already been deleted. Please refresh the folder.",
+		});
 };
